requirements.txt
streamlit
pandas
google-generativeai
python-dotenv
thefuzz
PyPDF2
plotly